homework#7 refactoring codes, have used disturization, taked functions to a separate file
