+++ index.html
@@ -10,7 +10,7 @@
 </head>
 <body>
   <div class="root">
-    <div class="arenas arena1">а если по
+    <div class="arenas arena3">а если по
       <form class="control">
         <div class="inputWrap">
           <h2>HIT</h2>
@@ -57,7 +57,6 @@
     <div class="wall-left"></div>
     <div class="wall-right"></div>
   </div>
-  <script src="logs.js"></script>
-  <script src="main.js"></script>
+  <script src="main.js" type="module"></script>
 </body>
 </html>
